unit test

+++ tracker-app/src/main/java/com/github/igordavydenko/tracker/rest/dto/StatisticDto.java
@@ -6,14 +6,13 @@ import lombok.Setter;
 
 import java.io.Serializable;
 
-
 @Getter
 @Setter
 @Builder
 public class StatisticDto implements Serializable {
 
   private int totalRuns;
-  private double totalDistance;
+  private int totalDistance;
   private double averageSpeed;
 
 }
